atof and helper functions in new file to make more sense.
