+++ src/utility/ostream.cc
@@ -97,24 +97,4 @@ int OStream::ptoa(const void * p, char * s)
     return j + 2;
 }
 
-// src https://searchcode.com/codesearch/view/10260157/
-bool OStream::isspace(unsigned char c) {
-    if ( c == ' '
-        || c == '\f'
-        || c == '\n'
-        || c == '\r'
-        || c == '\t'
-        || c == '\v' )
-        return true;
-
-    return false;
-}
-
-// src https://searchcode.com/codesearch/view/10200599/
-bool OStream::isdigit(unsigned char c) {
-    if ( c >= '0' && c <= '9' )
-        return true;
-
-    return false;
-}
 __END_UTIL
